Improve addcarry() and BaseX encoding (#68)

+++ include/CppCore/Encoding.h
@@ -213,18 +213,18 @@ namespace CppCore
       {
          assert(::strlen(alphabet) >= 2);
          uint8_t tbl[256];
-         size_t n = 0;
+         uint8_t n = 0;
          CppCore::clear(r);
          CppCore::clear(tbl);
          while (const char c = *alphabet++)
-            tbl[c] = (uint8_t)n++;
+            tbl[c] = n++;
          if (const char c = *input++)
          {
             *(uint8_t*)&r = tbl[c];
             while (const char c = *input++)
             {
                CppCore::umul(r, n, r);
-               CppCore::uadd(r, (size_t)tbl[c], r);
+               CppCore::uadd(r, tbl[c], r);
             }
          }
       }
@@ -239,10 +239,10 @@ namespace CppCore
          if (!input || !alphabet)
             return false; // null pointer
          uint8_t tbl[256];
-         size_t n = 0;
+         uint8_t n = 0;
          CppCore::bytedup(0xFF, tbl);
          while (const char c = *alphabet++)
-            tbl[c] = (uint8_t)n++;
+            tbl[c] = n++;
          if (n < 2U) CPPCORE_UNLIKELY
             return false; // alphabet too short
          if (const char first = *input++) CPPCORE_LIKELY
@@ -262,7 +262,7 @@ namespace CppCore
                if (t.of != 0U) CPPCORE_UNLIKELY
                   return false; // mul overflow
                uint8_t carry = 0;
-               CppCore::addcarry(t.v, (size_t)idx, r, carry);
+               CppCore::addcarry(t.v, idx, r, carry);
                if (carry != 0) CPPCORE_UNLIKELY
                   return false; // add overflow
             }
